Refactored sample project
