+++ samples/Sample.AspNetCore.SwaggerUI.Swashbuckle/Controllers/SampleController.cs
@@ -1,16 +1,18 @@
-﻿using Microsoft.AspNetCore.Authorization;
+﻿#pragma warning disable S1133 // Deprecated code should be removed - Done for testing purposes
+#pragma warning disable CS0618 // Type or member is obsolete - Done for testing purposes
+
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.ComponentModel.DataAnnotations;
 
 namespace Controllers;
 
-#pragma warning disable CS0618 // Type or member is obsolete - Done for test purpose
-
 [ApiController]
+[Route("sample")]
 public class SampleController : ControllerBase
 {
     [HttpHead("head")]
-    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(typeof(int), StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public IActionResult HeadSample([Required] int sample)
     {
@@ -18,9 +20,9 @@ public class SampleController : ControllerBase
     }
 
     [HttpPatch("patch")]
-    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(typeof(Models.Sample), StatusCodes.Status200OK)]
     [Produces("application/json")]
-    public IActionResult PatchSample(Sample model)
+    public IActionResult PatchSample(Models.Sample model)
     {
         return Ok(model);
     }
@@ -34,19 +36,16 @@ public class SampleController : ControllerBase
 
     [HttpPost("form")]
     [Authorize]
-    public IActionResult PostSample([FromForm] Sample model, IFormFile form)
+    public IActionResult PostSample([FromForm] Models.Sample model, IFormFile form)
     {
         return Ok();
     }
 
     [HttpGet("get")]
-#pragma warning disable S1133 // Deprecated code should be removed - Done for test purpose
+    [ProducesResponseType(typeof(int), StatusCodes.Status200OK)]
     [Obsolete("test")]
-#pragma warning restore S1133 // Deprecated code should be removed - Done for test purpose
     public IActionResult GetDisabledSample([Required] int sample)
     {
         return Ok(sample);
     }
 }
-
-#pragma warning restore CS0618 // Type or member is obsolete - Done for test purpose
